Moved to 'flat' project layout, adjusted local import paths
https://packaging.python.org/en/latest/discussions/src-layout-vs-flat-layout/#src-layout-vs-flat-layout
