+++ rivulet/riv_exceptions.py
